requirements.txt


requests
pyTelegramBotAPI
telebot-calendar
python-dotenv
aiohttp